Move the most general rule last
Per the advice in https://www.openldap.org/doc/admin24/access-control.html,
specifically section "8.4.10. When things aren't working as expected",
which notes that the access config stops searching when it finds
an access rule which matches (regardless of the grant).

The issue here therefore was that the 'to *' part was matching
_everything_ and the later rules weren't being checked.